module_4/app/requirements.txt
streamlit
psycopg2-binary
elasticsearch
openai
python-dotenv